Adding DB-API 2.0 (PEP-249) compliant cursor methods
Included are the following DB-API 2.0 compliant methods: fetchone(),
fetchmany() and fecthall(), as well as arraysize (default 1000).
These methods allow for standard access to Hive via HiveServer2. The
implementation is thread-safe, to the degree that the same cursor could
be used in multiple threads. However, I'm not sure that multiple cursors
may be used simultaneously.

These methods have not been thoroughly tested.